Cambiar temperatura de IntegerField a FloatField para soportar decimales
- Modelo Temperatura: valor ahora es FloatField (permite 34.5, 35.4, etc)
- Modelo ControlMonitoreo: temperatura ya era FloatField (sin cambios)
- Migración: 0005_alter_temperatura_valor.py
- Permite registrar temperaturas con decimal como 38.5°C, 39.2°C

+++ Backend/temp_car/models.py
@@ -110,7 +110,7 @@ class PersonalInfo(models.Model):
 class Temperatura(models.Model):
     """Modelo legacy de temperatura - se mantiene para compatibilidad"""
     id_Temperatura = models.AutoField(primary_key=True)
-    valor = models.IntegerField()
+    valor = models.FloatField()
 
     def __str__(self):
         return str(self.valor)
